+++ blocks/energy-scale/src/edit.js
@@ -1,0 +1,44 @@
+/**
+ * Add individual dependencies.
+ */
+import {
+	useBlockProps
+} from '@wordpress/block-editor';
+import ServerSideRender from '@wordpress/server-side-render';
+const { useEffect, useState } = wp.element;
+
+/**
+ * The edit function describes the structure of your block in the context of the
+ * editor. This represents what the editor will render when the block is used.
+ *
+ * @see https://developer.wordpress.org/block-editor/developers/block-api/block-edit-save/#edit
+ *
+ * @param object
+ * @return {WPElement} Element to render.
+ */
+export default function Edit( object ) {
+
+  const [ query, setQuery ] = useState( '' );
+
+	// secure ID of this block
+  useEffect( () => {
+    if ( ! object.attributes.blockId ) {
+      object.setAttributes( { blockId: object.clientId } );
+    }
+  }, [ object.attributes.blockId, object.clientId ] );
+
+  const isPreview = !! object.attributes.preview;
+
+	/**
+	 * Collect return for the edit-function
+	 */
+	return (
+		<div { ...useBlockProps() }>
+			<ServerSideRender
+				block="connector-for-propstack/energy-scale"
+				attributes={ object.attributes }
+				httpMethod="POST"
+			/>
+		</div>
+	);
+}